reduce options code duplication
