+++ packages/dds/sequence/src/test/fuzz/fuzzUtils.ts
@@ -3,13 +3,18 @@
  * Licensed under the MIT License.
  */
 
+import * as path from "path";
 import { strict as assert } from "assert";
 import {
 	AcceptanceCondition,
 	combineReducersAsync as combineReducers,
 	AsyncReducer as Reducer,
 } from "@fluid-internal/stochastic-test-utils";
-import { DDSFuzzModel, DDSFuzzTestState } from "@fluid-internal/test-dds-utils";
+import {
+	DDSFuzzModel,
+	DDSFuzzSuiteOptions,
+	DDSFuzzTestState,
+} from "@fluid-internal/test-dds-utils";
 import {
 	IChannelAttributes,
 	IChannelServices,
@@ -375,4 +380,28 @@ export const baseModel: Omit<
 			}
 		},
 	],
+};
+
+export const defaultFuzzOptions: Partial<DDSFuzzSuiteOptions> = {
+	validationStrategy: { type: "fixedInterval", interval: 10 },
+	reconnectProbability: 0.1,
+	numberOfClients: 3,
+	clientJoinOptions: {
+		maxNumberOfClients: 6,
+		clientAddProbability: 0.1,
+	},
+	defaultTestCount: 100,
+	saveFailures: { directory: path.join(__dirname, "../../src/test/results") },
+	parseOperations: (serialized: string) => {
+		const operations: Operation[] = JSON.parse(serialized);
+		// Replace this value with some other interval ID and uncomment to filter replay of the test
+		// suite to only include interval operations with this ID.
+		// const filterIntervalId = "00000000-0000-0000-0000-000000000000";
+		// if (filterIntervalId) {
+		// 	return operations.filter((entry) =>
+		// 		[undefined, filterIntervalId].includes((entry as any).id),
+		// 	);
+		// }
+		return operations;
+	},
 };
